XW-474 | make code more readable
